# schedule-service/docs/swagger.yaml
basePath: /raspyx/api
definitions:
...
  models.AddSubjectTypeRequest:
    properties:
      type:
        example: Лекция
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  models.ResponseAPI:
    properties:
      errors:
        $ref: '#/definitions/models.ErrorResponse'
      message:
        type: string
      requestId:
        type: string
      result: {}
      success:
        type: boolean
    type: object
...
  models.SubjectType:
    properties:
      type:
        example: Лекция
        type: string
      uuid:
        example: 4b6c34bd-01f0-4fbe-be65-217984b3e33d
        type: string
    type: object
  models.SubjectTypeResponse:
    properties:
      subject_type_uuid:
        example: 4b6c34bd-01f0-4fbe-be65-217984b3e33d
        type: string
    type: object
  models.TeacherResponse:
    properties:
      teacher_uuid:
        example: 4b6c34bd-01f0-4fbe-be65-217984b3e33d
        type: string
    type: object
...
  models.UpdateSubjectTypeRequest:
    properties:
      type:
        example: Лекция
        type: string
    type: object
  models.UpdateTeacherRequest:
    properties:
      first_name:
        example: Имя
        type: string
      middle_name:
        example: Отчество
        type: string
      second_name:
        example: Фамилия
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for schedules
  title: Raspyx
  version: 2.0.0
paths:
...
  /v2/subject_types:
    get:
      consumes:
      - application/json
      description: Получение списка всех типов дисциплины в системе
      produces:
      - application/json
      responses:
        "200":
          description: Данные о типах дисциплины успешно получены
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseAPI'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/models.SubjectType'
                  type: array
              type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ResponseAPI'
      security:
      - ApiKeyAuth: []
      summary: Получение типов дисциплины
      tags:
      - subject_types
    post:
      consumes:
      - application/json
      description: Создание нового типа дисциплины
      parameters:
      - description: Данные нового типа дисциплины
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AddSubjectTypeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Тип дисциплины успешно создан
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseAPI'
            - properties:
                result:
                  $ref: '#/definitions/models.SubjectTypeResponse'
              type: object
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "409":
          description: Конфликт при создании типа дисциплины
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ResponseAPI'
      security:
      - ApiKeyAuth: []
      summary: Добавление типа дисциплины
      tags:
      - subject_types
  /v2/subject_types/{subject_type_uuid}:
    delete:
      consumes:
      - application/json
      description: Удаление типа дисциплины из системы по его идентификатору
      parameters:
      - description: Идентификатор типа дисциплины
        in: path
        name: subject_type_uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Тип дисциплины успешно удален
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseAPI'
            - properties:
                result:
                  $ref: '#/definitions/models.SubjectTypeResponse'
              type: object
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "404":
          description: Тип дисциплины не найдена
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "409":
          description: Конфликт при удалении типа дисциплины
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ResponseAPI'
      security:
      - ApiKeyAuth: []
      summary: Удаление типа дисциплины
      tags:
      - subject_types
    put:
      consumes:
      - application/json
      description: Обновление информации о типе дисциплины по его идентификатору
      parameters:
      - description: Идентификатор типа дисциплины
        in: path
        name: subject_type_uuid
        required: true
        type: string
      - description: Данные для обновления типа дисциплины
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateSubjectTypeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Данные о типе дисциплины успешно обновлены
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseAPI'
            - properties:
                result:
                  $ref: '#/definitions/models.SubjectTypeResponse'
              type: object
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "404":
          description: Тип дисциплины не найден
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "409":
          description: Конфликт при обновлении типа дисциплины
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ResponseAPI'
      security:
      - ApiKeyAuth: []
      summary: Обновление типа дисциплины
      tags:
      - subject_types
  /v2/subject_types/type/{subject_type}:
    get:
      consumes:
      - application/json
      description: Получение информации о типе дисциплины по его subject_type
      parameters:
      - description: Название типа дисциплины
        in: path
        name: subject_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Данные о типе дисциплины успешно получены
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseAPI'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/models.SubjectType'
                  type: array
              type: object
        "404":
          description: Тип дисциплины не найден
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ResponseAPI'
      security:
      - ApiKeyAuth: []
      summary: Получение типа дисциплины по его типу
      tags:
      - subject_types
  /v2/subject_types/uuid/{subject_type_uuid}:
    get:
      consumes:
      - application/json
      description: Получение информации о типе дисциплины по его subject_type_uuid
      parameters:
      - description: Идентификатор типа дисциплины
        in: path
        name: subject_type_uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Данные о типе дисциплины успешно получены
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseAPI'
            - properties:
                result:
                  $ref: '#/definitions/models.SubjectType'
              type: object
        "404":
          description: Тип дисциплины не найден
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ResponseAPI'
      security:
      - ApiKeyAuth: []
      summary: Получение типа дисциплины по идентификатору
      tags:
      - subject_types
...
  /v2/teachers/{teacher_uuid}:
    delete:
      consumes:
      - application/json
      description: Удаление преподавателя из системы по его идентификатору
      parameters:
      - description: Идентификатор преподавателя
        in: path
        name: teacher_uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Преподаватель успешно удалена
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseAPI'
            - properties:
                result:
                  $ref: '#/definitions/models.TeacherResponse'
              type: object
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "404":
          description: Преподаватель не найден
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "409":
          description: Конфликт при удалении преподавателя
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ResponseAPI'
      security:
      - ApiKeyAuth: []
      summary: Удаление преподавателя
      tags:
      - teachers
    put:
      consumes:
      - application/json
      description: Обновление информации о преподавателе по его идентификатору
      parameters:
      - description: Идентификатор преподавателя
        in: path
        name: teacher_uuid
        required: true
        type: string
      - description: Данные для обновления преподавателя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTeacherRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Данные о преподавателе успешно обновлены
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseAPI'
            - properties:
                result:
                  $ref: '#/definitions/models.TeacherResponse'
              type: object
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "404":
          description: Преподаватель не найден
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "409":
          description: Конфликт при обновлении преподавателя
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ResponseAPI'
      security:
      - ApiKeyAuth: []
      summary: Обновление преподавателя
      tags:
      - teachers
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"